allign correctly the leading edge when scrolling scene map carousel

+++ MovieWanderer/View/LeftAlignedFlowLayout.swift
@@ -10,7 +10,7 @@ import UIKit
 
 struct Constants {
     struct ScenesCollection {
-        static let cellWidth: CGFloat = (UIScreen.main.bounds.width / 4) * 3 
+        static let cellWidth: CGFloat = CGFloat(Int((UIScreen.main.bounds.width / 4) * 3))
         static let lineSpacing: CGFloat = 10
     }
